Fix first-frame issue using SSR:false and localStorage
